Split button to save canvas as a specific component
