+++ src/components/CanvasSaveButton/CanvasSaveButton.scss
@@ -1,0 +1,4 @@
+.tm-share-button,
+.tm-share-button:hover {
+  color: white;
+}
