+++ src/components/CanvasSaveButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './CanvasSaveButton';
