+++ src/components/ShareMenu/ShareMenu.js
@@ -4,7 +4,7 @@ import OLMap from 'ol/Map';
 import { TiImage, TiSocialFacebook, TiSocialTwitter } from 'react-icons/ti';
 import { FiMail } from 'react-icons/fi';
 import BlankLink from '../BlankLink';
-import Button from '../Button';
+import CanvasSaveButton from '../CanvasSaveButton';
 
 const configPropType = PropTypes.arrayOf(
   PropTypes.shape({
@@ -59,79 +59,9 @@ const defaultProps = {
 };
 
 class ShareMenu extends PureComponent {
-  static getDownloadImageName() {
-    return window.document.title.replace(/ /g, '_').toLowerCase();
-  }
-
-  getCanvasImage(opts, asMSBlob) {
-    const { map } = this.props;
-    let image;
-
-    map.once('postcompose', evt => {
-      const { canvas } = evt.context;
-      const { height, width } = canvas;
-
-      const destinationCanvas = document.createElement('canvas');
-      destinationCanvas.width = width;
-      destinationCanvas.height = height;
-
-      const destContext = destinationCanvas.getContext('2d');
-      destContext.fillStyle = 'white';
-      destContext.fillRect(0, 0, width, height);
-      destContext.drawImage(canvas, 0, 0, width, height, 0, 0, width, height);
-
-      if (asMSBlob) {
-        image = destinationCanvas.msToBlob();
-      } else {
-        image = destinationCanvas.toDataURL(opts.format || 'image/jpeg');
-      }
-    });
-    map.renderSync();
-    return image;
-  }
-
-  downloadCanvasImage(e, options) {
-    const opts = options || { format: 'image/jpeg' };
-    if (/msie (9|10)/gi.test(window.navigator.userAgent.toLowerCase())) {
-      // ie 9 and 10
-      const w = window.open('about:blank', '');
-      w.document.write(
-        `<img src="${this.getCanvasImage(opts)}" alt="from canvas"/>`,
-      );
-    } else if (window.navigator.msSaveBlob) {
-      // ie 11 and higher
-      window.navigator.msSaveBlob(
-        new Blob([this.getCanvasImage(opts, true)], { type: opts.format }),
-        `${ShareMenu.getDownloadImageName()}.jpg`,
-      );
-    } else {
-      const link = document.createElement('a');
-      link.download = `${ShareMenu.getDownloadImageName()}.jpg`;
-      link.href = this.getCanvasImage(opts);
-      link.click();
-    }
-
-    if (window.navigator.msSaveBlob) {
-      // ie only
-      e.preventDefault();
-      e.stopPropagation();
-    }
-  }
-
-  renderImgSaveButton(conf) {
-    return (
-      <Button
-        className="tm-share-button"
-        title="Karte als Bild speichern."
-        onClick={e => this.downloadCanvasImage(e)}
-      >
-        <span>{conf.icon}</span>
-      </Button>
-    );
-  }
-
   render() {
     const {
+      map,
       className,
       classNameIcon,
       socialShareConfig,
@@ -159,7 +89,7 @@ class ShareMenu extends PureComponent {
                 {conf.icon}
               </BlankLink>
             ) : (
-              this.renderImgSaveButton(conf)
+              <CanvasSaveButton conf={conf} map={map} />
             )}
           </div>
         ))}
