+++ src/components/ShareMenu/ShareMenu.scss
@@ -41,8 +41,6 @@
     }
 
     a,
-    .tm-share-button,
-    .tm-share-button:hover,
     a:hover {
       color: white;
     }
