+++ src/components/ShareMenu/ShareMenu.test.js
@@ -1,3 +1,4 @@
+import 'jest-canvas-mock';
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { configure } from 'enzyme';
