+++ src/themes/default/components.scss
@@ -5,6 +5,7 @@
 @import '../../components/Button/Button.scss';
 @import '../../components/Footer/Footer.scss';
 @import '../../components/ShareMenu/ShareMenu.scss';
+@import '../../components/CanvasSaveButton/CanvasSaveButton.scss';
 @import '../../components/Menu/Menu.scss';
 @import '../../components/MenuItem/MenuItem.scss';
 @import '../../components/ToggleButton/ToggleButton.scss';
